+++ packages/components/src/components/model-signature/model-signature-styles.ts
@@ -9,8 +9,17 @@ import {
   type ModelSignatureSize,
 } from './model-signature-utils';
 
+/**
+ * @css-variable {"name": "--p-model-signature-width", "description": "Defines the width of the model signature. Overrides the `size` property when set.", "defaultValue": ""}
+ */
 const cssVariableWidth = '--p-model-signature-width';
+/**
+ * @css-variable {"name": "--p-model-signature-height", "description": "Defines the height of the model signature. Overrides the `size` property when set.", "defaultValue": ""}
+ */
 const cssVariableHeight = '--p-model-signature-height';
+/**
+ * @css-variable {"name": "--p-model-signature-color", "description": "Defines the color of the model signature. Overrides the `color` property when set.", "defaultValue": ""}
+ */
 const cssVariableColor = '--p-model-signature-color';
 
 const colorMap: Record<ModelSignatureColor, string> = {
@@ -28,7 +37,6 @@ export const getComponentCss = (
   color: ModelSignatureColor
 ): string => {
   const { width, height } = MODEL_SIGNATURES_MANIFEST[model];
-  const isSizeInherit = size === 'inherit';
 
   return getCss({
     '@global': {
@@ -37,10 +45,12 @@ export const getComponentCss = (
         verticalAlign: 'top',
         maxWidth: '100%',
         maxHeight: '100%',
-        // width + height style can't be !important atm to be backwards compatible with e.g. `<p-model-signature size="inherit" style="height: 50px"/>`
-        width: `var(${cssVariableWidth},${isSizeInherit ? 'auto' : `${width}px`})`,
-        height: `var(${cssVariableHeight},auto)`,
+
         ...addImportantToEachRule({
+          ...(size !== 'inherit' && {
+            height: `var(${cssVariableHeight}, auto)`,
+            width: `var(${cssVariableWidth}, ${width}px)`,
+          }),
           mask: `url(${getSvgUrl(model)}) no-repeat left top / contain`,
           aspectRatio: `${width} / ${safeZone ? 36 : height}`, // 36px is the max-height for SVG model signature creation
           background: `var(${cssVariableColor},${colorMap[color]})`, // color="inherit" will use currentcolor for inheritance