+++ packages/components/CHANGELOG.md
@@ -58,6 +58,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 - `Ai Tag`: New component to mark AI-generated/AI-modified content
   ([#4400](https://github.com/porsche-design-system/porsche-design-system/pull/4400))
 
+### Changed
+
+- `Wordmark`, `Model Signature`: Deprecate `size` and allow custom sizing via CSS Variables
+  ([#4393](https://github.com/porsche-design-system/porsche-design-system/pull/4393))
+
 ## [4.1.0] - 2026-05-06
 
 ## [4.1.0-rc.0] - 2026-05-05